pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "siqe-v3"
version = "3.0.0"
description = "Self-Improving Quant Engine V3"
authors = [{name = "SIQE Team"}]
dependencies = [
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "duckdb>=1.5.1",
    "fastapi>=0.135.3",
    "uvicorn>=0.43.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "loguru>=0.7.3",
    "pytz>=2025.2",
    "prometheus_client>=0.20.0",
    "ccxt>=4.0.0",
    "pyarrow>=14.0.0",
]

[project.urls]
Homepage = "https://github.com/anomalyco/siqe-v3"

[project.scripts]
siqe = "siqe.main:main"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*", "scripts*", "backtest*", "output*", "logs*", "data*", "*.egg-info*", "build*", "dist*", ".venv", "venv"]